Add timeouts to component processors
Signal processors can take arbitrarilly long and may hang if there is
some failure that results in the promise never resolving.

This can make it hard to debug what components are failing both in
production and in development environments.

This change adds a processor timeout configuration option and gives it a
default of 10000ms.